Moving canActivate to common guardUtil
