+++ src/app/guards/acceptTerms.guard.spec.ts
@@ -3,14 +3,12 @@ import { AcceptTermsGuard } from './acceptTerms.guard';
 describe('Accept terms guard', () => {
     let guard: AcceptTermsGuard;
     let mockStore: any;
-    let mockService: any;
     let mockGuardUtil: any;
 
     beforeEach(() => {
         mockStore = jasmine.createSpyObj('mockStore', ['unsubscribe', 'dispatch', 'pipe']);
-        mockService = jasmine.createSpyObj('mockService', ['get']);
         mockGuardUtil = jasmine.createSpyObj('guardUtil', ['checkStore', 'handleTC']);
-        guard = new AcceptTermsGuard(mockStore, mockService, mockGuardUtil);
+        guard = new AcceptTermsGuard(mockStore, mockGuardUtil);
     });
 
     it('is Truthy', () => {
