+++ src/app/guards/allowAcceptTerms.guard.spec.ts
@@ -3,14 +3,12 @@ import { AllowAcceptTermsGuard } from './allowAcceptTerms.guard';
 describe('Allow Accept Terms guard', () => {
     let guard: AllowAcceptTermsGuard;
     let mockStore: any;
-    let mockService: any;
     let mockGuardUtil: any;
 
     beforeEach(() => {
         mockStore = jasmine.createSpyObj('mockStore', ['unsubscribe', 'dispatch', 'pipe']);
-        mockService = jasmine.createSpyObj('mockService', ['get']);
         mockGuardUtil = jasmine.createSpyObj('guardUtil', ['checkStore', 'handleTC']);
-        guard = new AllowAcceptTermsGuard(mockStore, mockService, mockGuardUtil);
+        guard = new AllowAcceptTermsGuard(mockStore, mockGuardUtil);
     });
 
     it('is Truthy', () => {
